+++ client/components/user-home/index.js
@@ -1,15 +1,15 @@
 export {default} from './user-home'
 
-export {default as WelcomeUser} from './user-comps/welcome-user'
+export {default as WelcomeUser} from './welcome-user'
 
 //user
 export {default as Orders} from './user-comps/orders'
 export {default as Reviews} from './user-comps/reviews'
 
 //admin
-export {default as UserList} from './user-comps/user-list'
-export {default as OrderList} from './user-comps/order-list'
-export {default as ProductList} from './user-comps/product-list'
+export {default as UserList} from './user-admin-comps/user-list'
+export {default as OrderList} from './user-admin-comps/order-list'
+export {default as ProductList} from './user-admin-comps/product-list'
 
 //forms
-export {default as UserUpdate} from './user-forms/user-update'
+export {default as UserUpdate} from './user-actions/user-update'
