+++ client/components/user-home/welcome-user.js
